Segregate login in Auth module
