+++ backend/.env
@@ -1,1 +1,2 @@
 DATABASE_URL="postgresql://admin:123456@postgres:5432/nintendin"
+JWT_SECRET="4399896438a018e09ef8460ba1b4481e8d9fbe4800f20801de00e22e1d846897f6f9a9ea6494919e7f870ed283bdfcf45242604040d4fbdd4a43b2079669abb1"
