+++ backend/src/modules/auth/auth.controller.js
@@ -1,0 +1,24 @@
+import UserService from "./user.service.js";
+
+const AuthController = {};
+
+AuthController.loginUser = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+    const user = await UserService.findByEmailAndPassword(email, password);
+    if (!user) {
+      return res.status(404).json({ message: "Usuário e senha incorretos." });
+    }
+    const token = jwt.sign(
+      { id: user.id, email: user.email }, 
+      process.env.JWT_SECRET, 
+      { expiresIn: 3600 } 
+    );
+    res.status(200).json({ message: "Login realizado com sucesso!", token, user });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json("Internal server error");
+  }
+};
+
+export default AuthController;
