+++ backend/src/modules/auth/auth.service.js
@@ -1,0 +1,5 @@
+
+
+const AuthService = {};
+
+export default AuthService;
