+++ backend/src/modules/user/user.controller.js
@@ -2,7 +2,6 @@ import UserService from "./user.service.js";
 import jwt from "jsonwebtoken";
 
 const UserController = {};
-const SECRET="Teste123";
 // get, post, put, patch, etc ...
 UserController.getUser = async (req, res) => {
   try {
@@ -24,25 +23,6 @@ UserController.insertUser = async (req, res) => {
     };
     await UserService.create(body);
     res.status(200).json("User created successfully");
-  } catch (error) {
-    console.error(error);
-    res.status(500).json("Internal server error");
-  }
-};
-UserController.loginUser = async (req, res) => {
-  try {
-    const { email, password } = req.body;
-    const user = await UserService.findByEmailAndPassword(email, password);
-
-    if (!user) {
-      return res.status(404).json({ message: "Usuário e senha Incorrétos." });
-    }
-    const token = jwt.sign(
-      { id: user.id, email: user.email }, 
-      SECRET, 
-      { expiresIn: 3600 } 
-    );
-    res.status(200).json({ message: "Login realizado com sucesso! " + "Token: "+ token, user });
   } catch (error) {
     console.error(error);
     res.status(500).json("Internal server error");
