+++ backend/src/modules/user/user.route.js
@@ -6,7 +6,6 @@ const UserRouter = express.Router();
 // Define your routes here
 UserRouter.get("/", UserController.getUser);
 UserRouter.post("/", UserController.insertUser);
-UserRouter.post("/login", UserController.loginUser);
 UserRouter.put("/:id", UserController.updateUser);
 UserRouter.delete("/:id", UserController.deleteUser);
